Added test script and some optimizations

+++ Assets/Scripts/DNA.cs
@@ -7,9 +7,9 @@ public class DNA<T>
 
     private Random random;
     private Func<T> getRandomGene;
-    private Func<float, int> fitnessFunction;
+    private Func<int, float> fitnessFunction;
 
-    public DNA(int size, Random random, Func<T> getRandomGene, Func<float, int> fitnessFunction, bool shouldInitGenes = true)
+    public DNA(int size, Random random, Func<T> getRandomGene, Func<int, float> fitnessFunction, bool shouldInitGenes = true)
     {
         Genes = new T[size];
         this.random = random;
